Add standard middleware to handle missing routes
Each framework offers a different request handler or middleware to be
used as last item in the middleware pipeline, which is called when no
route has been matched by the router.

This allows us to delegate the formatting of such error to an error
handling middleware, granting more flexibility to people.